Some type safety conversions

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/detection/distance/VL53L1X_Version.java
@@ -1,34 +1,34 @@
 package org.firstinspires.ftc.teamcode.hardware.detection.distance;
 
 public class VL53L1X_Version {
-  public VL53L1X_Version(short major, short minor, short build, long revision) {
+  public VL53L1X_Version(byte major, byte minor, byte build, byte revision) {
     this.major = major;
     this.minor = minor;
     this.build = build;
     this.revision = revision;
   }
 
-  public short getMajor() {
+  public byte getMajor() {
     return major;
   }
 
-  public void setMajor(short major) {
+  public void setMajor(byte major) {
     this.major = major;
   }
 
-  public short getMinor() {
+  public byte getMinor() {
     return minor;
   }
 
-  public void setMinor(short minor) {
+  public void setMinor(byte minor) {
     this.minor = minor;
   }
 
-  public short getBuild() {
+  public byte getBuild() {
     return build;
   }
 
-  public void setBuild(short build) {
+  public void setBuild(byte build) {
     this.build = build;
   }
 
@@ -36,12 +36,12 @@ public class VL53L1X_Version {
     return revision;
   }
 
-  public void setRevision(long revision) {
+  public void setRevision(byte revision) {
     this.revision = revision;
   }
 
-  private short major;
-  private short minor;
-  private short build;
-  private long revision;
+  private byte major;
+  private byte minor;
+  private byte build;
+  private byte revision;
 }
